+++ src/basic_mpc/data/__init__.py
@@ -1,7 +1,7 @@
 """Chargement, modèle de capteur et prétraitement."""
 
 from basic_mpc.data.loading import load_raw_csv
-from basic_mpc.data.pipeline import regularize_series, run_preprocess, sampling_report
+from basic_mpc.data.pipeline import regularize_series, run_preprocess, sampling_report, to_indexed_series
 from basic_mpc.data.sensors import SensorModel, infer_resolution
 
 __all__ = [
@@ -11,4 +11,5 @@ __all__ = [
     "regularize_series",
     "run_preprocess",
     "sampling_report",
+    "to_indexed_series",
 ]